Mod: 收敛设置hexo服务器配置信息的逻辑
